Move storage adapters under plugin package
