+++ pkgs/standards/peagen/peagen/plugins/storage_adapters/github_storage_adapter.py
@@ -1,0 +1,6 @@
+class GithubStorageAdapter:
+    def __init__(self, **kwargs):
+        self.kwargs = kwargs
+
+    def upload(self, path: str, dest: str) -> str:
+        return f"github://{dest}"
