+++ pkgs/standards/peagen/pyproject.toml
@@ -132,10 +132,10 @@ init-doe-spec                         = "peagen.template_scaffold.doe_spec"
 init-ci                               = "peagen.template_scaffold.ci"
 
 [project.entry-points."peagen.storage_adapters"]
-minio   = "peagen.storage_adapters.minio_storage_adapter:MinioStorageAdapter"
-file    = "peagen.storage_adapters.file_storage_adapter:FileStorageAdapter"
-github    = "peagen.storage_adapters.github_storage_adapter:GithubStorageAdapter"
-gh_release    = "peagen.storage_adapters.gh_release_storage_adapter:GithubReleaseStorageAdapter"
+minio   = "peagen.plugins.storage_adapters.minio_storage_adapter:MinioStorageAdapter"
+file    = "peagen.plugins.storage_adapters.file_storage_adapter:FileStorageAdapter"
+github    = "peagen.plugins.storage_adapters.github_storage_adapter:GithubStorageAdapter"
+gh_release    = "peagen.plugins.storage_adapters.gh_release_storage_adapter:GithubReleaseStorageAdapter"
 
 [project.entry-points."peagen.publishers"]
 redis   = "peagen.plugins.publishers.redis_publisher:RedisPublisher"
